Updated functions
In HKF, started to add the functionality to automatically deal with x.data.  In HKF_MLE, added the option to use optim, which computes the hessian matrix - this can be used to derived standard errors. Also added, HKF_tests, here I compare the HKF package to the results of DLM. The test case in the local level model (random walk) applied to the Nile data set. Testing shows that the HKF package is working well.